refactor: Move newsletter form into footer
