+++ frontend/components/common/Layout/index.tsx
@@ -1,4 +1,5 @@
 import { Box, Flex } from '@chakra-ui/react';
+import { GlobalSettings } from '@models/global-settings';
 import { Store, StoreListItem } from '@models/store';
 import Head from 'next/head';
 import * as React from 'react';
@@ -9,12 +10,14 @@ export interface LayoutProps {
    title: string;
    stores: StoreListItem[];
    currentStore: Store;
+   globalSettings: GlobalSettings;
 }
 
 export function Layout({
    title,
    stores,
    currentStore,
+   globalSettings,
    children,
 }: React.PropsWithChildren<LayoutProps>) {
    return (
@@ -26,7 +29,11 @@ export function Layout({
          <Flex direction="column">
             <Header menu={currentStore.header.menu} />
             {children}
-            <Footer currentStore={currentStore} stores={stores} />
+            <Footer
+               currentStore={currentStore}
+               stores={stores}
+               globalSettings={globalSettings}
+            />
          </Flex>
       </Box>
    );
