+++ frontend/components/product-list/ProductListView.tsx
@@ -1,20 +1,12 @@
 import { Flex, VStack } from '@chakra-ui/react';
-import {
-   PageContentWrapper,
-   SecondaryNavbar,
-   SecondaryNavbarItem,
-   SecondaryNavbarLink,
-} from '@components/common';
-import { ALGOLIA_API_KEY, ALGOLIA_APP_ID, IFIXIT_ORIGIN } from '@config/env';
+import { PageContentWrapper, SecondaryNavbar } from '@components/common';
+import { ALGOLIA_API_KEY, ALGOLIA_APP_ID } from '@config/env';
 import { AlgoliaProvider, SearchContext } from '@lib/algolia';
-import { GlobalSettings } from '@models/global-settings';
 import {
    ProductList,
    ProductListSectionType,
    ProductSearchHit,
 } from '@models/product-list';
-import snakeCase from 'lodash/snakeCase';
-import NextLink from 'next/link';
 import { MetaTags } from './MetaTags';
 import { ProductListBreadcrumb } from './ProductListBreadcrumb';
 import { ProductListDeviceNavigation } from './ProductListDeviceNavigation';
@@ -23,7 +15,6 @@ import {
    FeaturedProductListSection,
    FilterableProductsSection,
    HeroSection,
-   NewsletterSection,
    ProductListChildrenSection,
    ProductListSetSection,
    RelatedPostsSection,
@@ -32,16 +23,12 @@ import {
 export interface ProductListViewProps {
    productList: ProductList;
    searchContext: SearchContext<ProductSearchHit>;
-   globalSettings: GlobalSettings;
 }
 
 export function ProductListView({
    productList,
    searchContext,
-   globalSettings,
 }: ProductListViewProps) {
-   const { newsletterForm } = globalSettings;
-   const isRootProductList = productList.ancestors.length === 0;
    return (
       <>
          <SecondaryNavbar>
@@ -149,16 +136,6 @@ export function ProductListView({
                         }
                      }
                   })}
-                  {newsletterForm && (
-                     <NewsletterSection
-                        title={newsletterForm.title}
-                        description={newsletterForm.subtitle}
-                        emailPlaceholder={
-                           newsletterForm.inputPlaceholder || undefined
-                        }
-                        subscribeLabel={newsletterForm.callToActionButtonTitle}
-                     />
-                  )}
                </AlgoliaProvider>
             </VStack>
          </PageContentWrapper>
