+++ frontend/components/product-list/sections/index.ts
@@ -1,6 +1,5 @@
 export * from './BannerSection';
 export * from './RelatedPostsSection';
-export * from './NewsletterSection';
 export * from './FilterableProductsSection';
 export * from './HeroSection';
 export * from './ProductListChildrenSection';
